Fix 'comparison between signed and unsigned integer expressions' warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@31229 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gml/gmlhandler.cpp
@@ -1302,7 +1302,7 @@ OGRErr GMLHandler::startElementDefault(const char *pszName, int nLenName, void* 
     /* WFS 2.0 GetFeature documents have a wfs:FeatureCollection */
     /* as a wfs:member of the top wfs:FeatureCollection. We don't want this */
     /* wfs:FeatureCollection to be recognized as a feature */
-    else if( (!(nLenName == strlen("FeatureCollection") &&
+    else if( (!(nLenName == (int)strlen("FeatureCollection") &&
                 strcmp(pszName, "FeatureCollection") == 0)) &&
              (nClassIndex = m_poReader->GetFeatureElementIndex( pszName, nLenName, eAppSchemaType )) != -1 )
     {
